\name{abn-defunct}
\alias{abn-defunct}
\alias{search.hillclimber}
\title{Defunct functions and data in package \pkg{abn}}
\description{
The functions and data listed are defunct in  package \pkg{abn}. Alternative data and functions with similar
functionality are mentioned.}
\usage{
  search.hillclimber(...)
}
\seealso{
    \code{\link{searchHillClimber}}
}
% R documentation directory (show via RShowDoc("KEYWORDS")):
\keyword{internal}
