\name{abn-deprecated}
\alias{abn-deprecated}
\alias{fitabn}
\alias{plotabn}
\alias{searchHillclimber}
\alias{mostprobable}
\alias{buildscorecache}
\title{Deprecated functions and data in package \pkg{abn}}
\description{
The functions and data listed are deprecated and/or renamed and will be defunct 
in the near future. Alternative data and functions with similar
functionality are mentioned.}
\usage{
  fitabn(...)
  plotabn(...)
  searchHillclimber(...)
  mostprobable(...)
  buildscorecache(...)
}
\seealso{
  \code{\link{plotAbn}},  \code{\link{searchHillclimber}},  \code{\link{mostProbable}}, \code{\link{buildScoreCache}},
}
% R documentation directory (show via RShowDoc("KEYWORDS")):
\keyword{internal}
