%                               -*- Mode: Rd -*- 
% bn-internal.Rd --- 
% Author          : Fraser Lewis
% Created On      : 
% Last Modified By: Fraser Lewis
% Last Modified On: 
% Update Count    : 
% Status          : Unknown, Use with caution!
% 
\name{bn-internal}

\alias{makeintofactors}
\alias{makeintobinarymatrix}
\alias{trim}

\title{bn internal functions}

\description{
  These are functions internally called by other functions
  in the package \code{abn} and not meant to be called by the user.
}

\usage{

makeintofactors(data.df)
makeintobinarymatrix(dag.df)
trim(val,threshold) 
}

\author{
  Fraser Lewis \email{fraseriain.lewis@uzh.ch}.
}


\keyword{internal}
