% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_node_inla_glmm.R
\name{calc.node.inla.glmm}
\alias{calc.node.inla.glmm}
\title{Fit a given regression using INLA}
\usage{
calc.node.inla.glmm(
  child.loc = NULL,
  dag.m.loc = NULL,
  data.df.loc = NULL,
  data.dists.loc = NULL,
  ntrials.loc = NULL,
  exposure.loc = NULL,
  compute.fixed.loc = NULL,
  mean.intercept.loc = NULL,
  prec.intercept.loc = NULL,
  mean.loc = NULL,
  prec.loc = NULL,
  loggam.shape.loc = NULL,
  loggam.inv.scale.loc = NULL,
  verbose.loc = FALSE,
  nthreads = NULL
)
}
\arguments{
\item{child.loc}{index of current child node.}

\item{dag.m.loc}{dag as matrix.}

\item{data.df.loc}{data df,}

\item{data.dists.loc}{list of distributions.}

\item{ntrials.loc}{\code{rep(1,dim(data.df)[1])}.}

\item{exposure.loc}{\code{rep(1,dim(data.df)[1])}.}

\item{compute.fixed.loc}{TRUE.}

\item{mean.intercept.loc}{the prior mean for all the Gaussian additive terms for each node. INLA argument \code{control.fixed=list(mean.intercept=...)} and \code{control.fixed=list(mean=...)}.}

\item{prec.intercept.loc}{the prior precision for all the Gaussian additive term for each node. INLA argument \code{control.fixed=list(prec.intercept=...)} and \code{control.fixed=list(prec=...)}.}

\item{mean.loc}{the prior mean for all the Gaussian additive terms for each node. INLA argument \code{control.fixed=list(mean.intercept=...)} and \code{control.fixed=list(mean=...)}. Same as \code{mean.intercept.loc}.}

\item{prec.loc}{the prior precision for all the Gaussian additive term for each node. INLA argument \code{control.fixed=list(prec.intercept=...)} and \code{control.fixed=list(prec=...)}. Same as \code{prec.intercept.loc}.}

\item{loggam.shape.loc}{the shape parameter in the Gamma distribution prior for the precision in a Gaussian node. INLA argument \code{control.family=list(hyper = list(prec = list(prior="loggamma",param=c(loggam.shape, loggam.inv.scale))))}.}

\item{loggam.inv.scale.loc}{the inverse scale parameter in the Gamma distribution prior for the precision in a Gaussian node. INLA argument \code{control.family=list(hyper = list(prec = list(prior="loggamma",param=c(loggam.shape, loggam.inv.scale))))}.}

\item{verbose.loc}{FALSE to not print additional output.}

\item{nthreads}{number of threads to use for INLA. Default is \code{fit.control[["ncores"]]} or \code{build.control[["ncores"]]} which is the number of cores specified in \code{control} and defaults to 1.}
}
\value{
If INLA failed, FALSE or an error is returned. Otherwise, the direct output from INLA is returned.
}
\description{
Internal wrapper to INLA and are called from \code{fitAbn.bayes} and \code{buildScoreCache.bayes}.
}
\seealso{
Other Bayes:
\code{\link[=buildScoreCache]{buildScoreCache()}},
\code{\link[=calc.node.inla.glm]{calc.node.inla.glm()}},
\code{\link[=fitAbn]{fitAbn()}},
\code{\link[=getmarginals]{getmarginals()}}
}
\concept{Bayes}
\keyword{internal}
