% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abn-internal.R
\name{check.valid.buildControls}
\alias{check.valid.buildControls}
\title{Simple check on the control parameters}
\usage{
check.valid.buildControls(control, method = "bayes", verbose = FALSE)
}
\arguments{
\item{control}{list of control arguments with new parameters supplied to \code{\link{buildScoreCache}} or \code{\link{fitAbn}}.}

\item{method}{"bayes" or "mle" strategy from argument \code{method=...} in \code{\link{buildScoreCache}} or \code{\link{fitAbn}}. Defaults to "bayes".}

\item{verbose}{when TRUE additional information is printed. Defaults to FALSE.}
}
\value{
list with all control arguments with respect to the method but with new values.
}
\description{
Simple check on the control parameters
}
\keyword{internal}
