% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abn-internal.R
\name{check.valid.dag}
\alias{check.valid.dag}
\title{Set of simple commonsense validity checks on the directed acyclic graph definition matrix}
\usage{
check.valid.dag(
  dag = NULL,
  data.df = NULL,
  is.ban.matrix = FALSE,
  group.var = NULL
)
}
\arguments{
\item{dag}{Named square matrix or a formula statement specifying a directed acyclic graph.
If NULL an empty network is assumed.}

\item{data.df}{data frame with names corresponding to variable/node names.}

\item{is.ban.matrix}{Diagonals and cycles are not checked for ban-matrices. Defaults to FALSE.}

\item{group.var}{not yet implemented}
}
\value{
dag as named square matrix
}
\description{
Set of simple commonsense validity checks on the directed acyclic graph definition matrix
}
\keyword{internal}
