% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abn-internal.R
\name{check.valid.data}
\alias{check.valid.data}
\title{Set of simple commonsense validity checks on the data.df and data.dists arguments}
\usage{
check.valid.data(data.df = NULL, data.dists = NULL, group.var = NULL)
}
\arguments{
\item{data.df}{data frame with names corresponding to variable/node names.}

\item{data.dists}{list specifying each columns distribution type. Names correspond to column names and values must be one of "gaussian", "binomial", "poisson", "multinomial".}

\item{group.var}{not yet implemented}
}
\value{
list of indexes for each distribution type

a list of the indexes for each distribution type
}
\description{
Set of simple commonsense validity checks on the data.df and data.dists arguments
}
\keyword{internal}
