% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abn-internal.R
\name{check.valid.groups}
\alias{check.valid.groups}
\title{Simple check on the grouping variable}
\usage{
check.valid.groups(
  group.var = NULL,
  data.df = NULL,
  cor.vars = NULL,
  verbose = FALSE
)
}
\arguments{
\item{group.var}{Name of grouping variable.}

\item{data.df}{data frame of all variables including the variable specified in \code{group.var} as factor.}

\item{cor.vars}{Name(s) of variables to which the grouping should be applied to.}

\item{verbose}{when TRUE additional information is printed. Defaults to FALSE.}
}
\value{
list with data.df, indexes of variables to which the grouping should be applied to and the associated group for each observation as integer.
}
\description{
Simple check on the grouping variable
}
\keyword{internal}
