% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abn-internal.R
\name{check.valid.parents}
\alias{check.valid.parents}
\title{Set of simple checks on the given parent limits}
\usage{
check.valid.parents(data.df = NULL, max.parents = NULL, group.var = NULL)
}
\arguments{
\item{data.df}{data frame}

\item{max.parents}{single integer for one overall max parent limit.
A list with names corresponding to variable/column names of \code{data.df} and individual parent limits.
NULL for no parent limit restriction(s).}

\item{group.var}{not yet implemented}
}
\value{
numeric vector of max number of parents per variable
}
\description{
Set of simple checks on the given parent limits
}
\keyword{internal}
