% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abn-internal.R
\name{check.which.valid.nodes}
\alias{check.which.valid.nodes}
\title{Set of simple checks on the list given as parent limits}
\usage{
check.which.valid.nodes(data.df = NULL, which.nodes = NULL, group.var = NULL)
}
\arguments{
\item{data.df}{data frame}

\item{which.nodes}{a vector giving the column indices of the variables to be included, if ignored all variables are included.}

\item{group.var}{not yet implemented}
}
\value{
integer vector of column indexes of valid nodes in data.df
}
\description{
Set of simple checks on the list given as parent limits
}
\keyword{internal}
