% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictABN.R
\name{check_evidence}
\alias{check_evidence}
\title{Checking the evidence}
\usage{
check_evidence(data, mydists, hypothesis, evidence)
}
\arguments{
\item{data}{A data frame containing the data (samples in rows, variables in columns).}

\item{mydists}{A list containing the distributions of the nodes of the graph.}

\item{hypothesis}{Node to predict.}

\item{evidence}{Known nodes that are used to predict the hypothesis.}
}
\value{
A list containing the evidence in the right format
}
\description{
Check the format of the evidence
}
\examples{
# load a data set
data <- ex1.dag.data 

# define the distributions of the node
mydists <- list(b1="binomial", 
p1="poisson", 
g1="gaussian", 
b2="binomial", 
p2="poisson", 
b3="binomial", 
g2="gaussian", 
b4="binomial", 
b5="binomial", 
g3="gaussian") 

hypothesis <- "g2"
evidence <- list("b2" = "y", b5 = "n")

evidence <- check_evidence(data, mydists, hypothesis, evidence)

}
