% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abn-toolbox.R
\name{compareEG}
\alias{compareEG}
\title{Compare two DAGs or essential graphs (EGs)}
\usage{
compareEG(ref, test)
}
\arguments{
\item{ref}{a matrix or a formula statement (see details for format) defining
the reference network structure, a directed acyclic graph (DAG).
Note that row names must be set or given in \code{node.names}
if the DAG is given via a formula statement.}

\item{test}{a matrix or a formula statement (see details for format) defining
the test network structure, a directed acyclic graph (DAG).
Note that row names must be set or given in \code{node.names}
if the DAG is given via a formula statement.}
}
\value{
See \code{\link{compareDag}}.
}
\description{
Wrapper around \code{\link{compareDag}} with \code{checkDAG = FALSE}, so that
essential graphs, whose undirected edges are coded symmetrically, can be
compared as well.
}
\seealso{
\code{\link{compareDag}}, \code{\link{essentialGraph}}
}
