\name{ex3.dag.data}
\docType{data}
\alias{ex3.dag.data}

\title{Validation data set for use with abn library examples}
\description{1000 observations across with 13 binary variables and one grouping variable. Real (anonymised) data of unknown structure. 
}
\usage{ex3.dag.data} 
\format{A data frame with 14 columns, where     
     \code{b1,b2,\dots,b13} are binary variables encoded as factors and
     \code{group} is a factor with 100 factors  defining the sampling
     groups (10 observations each).}
\keyword{datasets}
