\name{ex4.dag.data}
\docType{data}
\alias{ex4.dag.data}
     
\title{Valdiation data set for use with abn library examples}
\description{2000 observations across with 10 binary variables and one grouping variable. Real (anonymised) data of unknown structure.  
     }
\usage{ex4.dag.data} 
\format{A data frame with eleven columns:
    \code{group} factor with 85 levels defining sampling groups; 
     \code{b1,\dots,b10} binary variables encoded as factors.
     }
     

\keyword{datasets}
