\name{ex5.dag.data}
     \docType{data}
     \alias{ex5.dag.data}
     
     \title{Data set for use with abn library examples}
     \description{434 observations across with 18 variables, 6 binary and 12 continuous, and one grouping variable. Real (anonymised) data of unknown structure. 
     }
     \usage{ex5.dag.data} 
\format{A data frame, binary variables are factors
     \describe{
     \item{b1}{binary}
     \item{b2}{binary}
\item{b3}{binary}
\item{b4}{binary}
\item{b5}{binary}
\item{b6}{binary}
\item{g1}{continuous}
\item{g2}{continuous}
\item{g3}{continuous}
\item{g4}{continuous}
\item{g5}{continuous}
\item{g6}{continuous}
\item{g7}{continuous}
\item{g8}{continuous}
\item{g9}{continuous}
\item{g10}{continuous}
\item{g11}{continuous}
\item{g12}{continuous}
\item{group}{binary, defines sampling groups}
     }}


\keyword{datasets}
