\name{ex6.dag.data}
\docType{data}
\alias{ex6.dag.data}
     
\title{Valdiation data set for use with abn library examples}
\description{800 observations across with 8 variables, 1 count, 2 binary and 4 continuous, and 1 grouping variable. Real (anonymised) data of unknown structure. }
\usage{ex6.dag.data} 
\format{A data frame with eight columns. Binary variables are factors
  \describe{
    \item{p1}{count}
    \item{g1}{continuous}
    \item{g2}{continuous}
    \item{b1}{binary}
    \item{b2}{binary}
    \item{g3}{continuous}
    \item{g4}{continuous}
    \item{group}{factor,defines sampling groups}
}}


\keyword{datasets}
