% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abn-internal.R
\name{.onAttach}
\alias{.onAttach}
\title{Prints start up message}
\usage{
.onAttach(lib, pkg)
}
\value{
Prints startup message to the console
}
\description{
Prints start up message
}
\examples{
library(abn)
}
\keyword{internal}
