

\name{eightvar}
\alias{eightvar}
\non_function{}
\title{simulated dataset from an eight variable additive DAG}
\description{
 1000 observations simulated from an additive DAG of known structure. 
}
\format{A data frame with discrete variables 
  \describe{
    \item{X1}{Binary. Independent}
    \item{X2}{Binary. An additive function of X1 and X8}
    \item{X3}{Binary. An additive function of X1}
    \item{X4}{Binary. An additive function of X7}
    \item{X5}{Binary. An additive function of X1 and X2}
    \item{X6}{Binary. Independent}
    \item{X7}{Binary. Independent}
    \item{X8}{Binary. An additive function of X7}
    }
 }

\keyword{datasets}

