% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitabn_bayes.R
\name{eval.across.grid}
\alias{eval.across.grid}
\title{function to get marginal across an equal grid}
\usage{
eval.across.grid(mylist, n.grid, single)
}
\arguments{
\item{mylist}{list of matrices of two cols x, y}

\item{n.grid}{grid size}

\item{single}{NULL or TRUE if only a single node and parameter}
}
\value{
list
}
\description{
function to get marginal across an equal grid
}
\keyword{internal}
