% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex3.dag.data}
\alias{ex3.dag.data}
\title{Validation data set for use with abn library examples}
\format{
A data frame with 14 columns, where
\code{b1,b2,\dots,b13} are binary variables encoded as factors and
\code{group} is a factor with 100 factors  defining the sampling
groups (10 observations each).
}
\usage{
ex3.dag.data
}
\description{
1000 observations across with 13 binary variables and one grouping variable. Real (anonymised) data of unknown structure.
}
\keyword{datasets}
\keyword{internal}
