% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex4.dag.data}
\alias{ex4.dag.data}
\title{Valdiation data set for use with abn library examples}
\format{
A data frame with eleven columns:
\code{group} factor with 85 levels defining sampling groups;
\code{b1,\dots,b10} binary variables encoded as factors.
}
\usage{
ex4.dag.data
}
\description{
2000 observations across with 10 binary variables and one grouping variable. Real (anonymised) data of unknown structure.
}
\keyword{datasets}
\keyword{internal}
