% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex5.dag.data}
\alias{ex5.dag.data}
\title{Valdiation data set for use with abn library examples}
\format{
A data frame with 19 columns: \code{b1,\dots,b6} binary
variables, encoded as factors;
\code{g1,\dots,g12} continuous variables. Finally, the column
\code{group} defines sampling groups (encoded as a factor as well).
}
\usage{
ex5.dag.data
}
\description{
434 observations across with 18 variables, 6 binary and 12 continuous, and one grouping variable. Real (anonymised) data of unknown structure.
}
\keyword{datasets}
\keyword{internal}
