% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex6.dag.data}
\alias{ex6.dag.data}
\title{Valdiation data set for use with abn library examples}
\format{
A data frame with eight columns. Binary variables are factors
\describe{
\item{p1}{count}
\item{g1}{continuous}
\item{g2}{continuous}
\item{b1}{binary}
\item{b2}{binary}
\item{g3}{continuous}
\item{g4}{continuous}
\item{group}{factor,defines sampling groups}
}
}
\usage{
ex6.dag.data
}
\description{
800 observations across with 8 variables, 1 count, 2 binary and 4 continuous, and 1 grouping variable. Real (anonymised) data of unknown structure.
}
\keyword{datasets}
\keyword{internal}
