% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ex7.dag.data}
\alias{ex7.dag.data}
\title{Valdiation data set for use with abn library examples}
\format{
A data frame, binary variables are factors
\describe{
\item{b1}{binary}
\item{b2}{binary}
\item{group}{factor, defines sampling groups}
}
}
\usage{
ex7.dag.data
}
\description{
10648 observations across with 3 variables, 2 binary and 1 grouping variable.
Real (anonymised) data of unknown structure.
}
\keyword{datasets}
\keyword{internal}
