% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abn-toolbox.R
\name{expit}
\alias{expit}
\title{expit of proportions}
\usage{
expit(x)
}
\arguments{
\item{x}{numeric with values between \code{[0,1]}.}
}
\value{
numeric vector of same length as \code{x}.
}
\description{
See also the C implementation \code{?abn::expit_cpp()}.
}
\keyword{utilities}
