% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{expit_cpp}
\alias{expit_cpp}
\title{expit function}
\usage{
expit_cpp(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
a numeric vector
}
\description{
transform \code{x} either via the logit, or expit.
}
