% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_abnFit_mle_graph}
\alias{export_abnFit_mle_graph}
\title{Export graph metadata from abnFit objects fitted with MLE}
\usage{
export_abnFit_mle_graph(object, ...)
}
\arguments{
\item{object}{An object of class abnFit}

\item{...}{Additional export options}
}
\value{
A named list containing graph metadata.
}
\description{
Export graph metadata from abnFit objects fitted with MLE
}
\details{
This function extracts metadata about the graph from abnFit objects.
It retrieves information such as the fitting method, number of nodes, number of observations,
and model scores (AIC, BIC, log-likelihood). If a grouping variable was used during fitting,
it is also included.
}
\keyword{internal}
