% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_abnFit_mle_grouped_nodes}
\alias{export_abnFit_mle_grouped_nodes}
\title{Export node information from abnFit objects fitted with MLE (mixed effects)}
\usage{
export_abnFit_mle_grouped_nodes(object, ...)
}
\arguments{
\item{object}{An object of class abnFit fitted with method = "mle" and group.var specified.}

\item{...}{Additional arguments (currently unused)}
}
\value{
This function does not return a value. It raises an error.
}
\description{
Export node information from abnFit objects fitted with MLE (mixed effects)
}
\details{
This function is a placeholder for exporting node information from abnFit objects.
Currently, it raises an error indicating that export for grouped MLE models is not implemented.
}
\keyword{internal}
