% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{extract_parameters_by_distribution}
\alias{extract_parameters_by_distribution}
\title{Helper function to extract parameters based on distribution type}
\usage{
extract_parameters_by_distribution(
  coef_vec,
  se_vec,
  distribution,
  node_id,
  parent_nodes,
  start_counter,
  link_function
)
}
\description{
Helper function to extract parameters based on distribution type
}
\keyword{internal}
