% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{factorial_fast}
\alias{factorial_fast}
\title{Fast Iterative Reweighed Least Square algorithm for Poissons}
\usage{
factorial_fast(n)
}
\description{
IRLS to estimate network score of Poisson nodes.
}
\keyword{internal}
