% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abn-methods.R
\name{family.abnFit}
\alias{family.abnFit}
\title{Print family of objects of class \code{abnFit}}
\usage{
\method{family}{abnFit}(object, ...)
}
\arguments{
\item{object}{Object of class \code{abnFit}}

\item{...}{additional parameters. Not used at the moment.}
}
\value{
prints the distributions for each variable of the fitted model.
}
\description{
Print family of objects of class \code{abnFit}
}
