% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictABN.R
\name{find_parents}
\alias{find_parents}
\title{Find the parents of a node in a graph}
\usage{
find_parents(graph, node)
}
\arguments{
\item{graph}{A directed graph (igraph object).}

\item{node}{A node of the graph (either a node label or a number)}
}
\value{
A vector containing the parents of the node (labels of the nodes if the graph is labeled)
}
\description{
Find the parents of a given node in a given graph.
}
\examples{
g <- make_graph("Zachary") # undirected graph
find_parents(g, node = 1) # neighbors of the node 1
}
