% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abn-internal.R
\name{formula_abn}
\alias{formula_abn}
\title{Formula to adjacency matrix}
\usage{
formula_abn(f, name)
}
\value{
A square matrix
}
\description{
Internal function that produce a square matrix length(name) with \eqn{0,1} depending on f.
f have to start with ~ terms are entries of name terms are separated by + term1 | term2 indicates
col(term1) row(term2) puts a 1 term1 | term2:term3: ... : is used as a sep . = all terms in name
}
\keyword{internal}
