% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{g2b2c_data}
\alias{g2b2c_data}
\title{Toy Data Set for Examples in README}
\format{
A data frame with five columns. Binary and categorical variables are factors.
\describe{
\item{G1}{gaussian}
\item{B1}{binomial}
\item{B2}{binomial}
\item{C}{categorical}
\item{G2}{gaussian}
}
}
\usage{
g2b2c_data
}
\description{
1000 observations with 5 variables: 2 continuous, 2 binary and 1 categorical.
}
\keyword{datasets}
\keyword{internal}
