% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{g2pbcgrp}
\alias{g2pbcgrp}
\title{Toy Data Set for Examples in README}
\format{
A data frame with six columns. Binary and categorical variables are factors.
\describe{
\item{G1}{gaussian}
\item{P}{poisson}
\item{B}{binomial}
\item{C}{categorical}
\item{G2}{gaussian}
\item{group}{categorical}
}
}
\usage{
g2pbcgrp
}
\description{
10000 observations with 6 variables: 2 continuous, 1 binary, 1 count, 1 categorical and 1 grouping factor.
}
\keyword{datasets}
\keyword{internal}
