% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitabn_bayes.R
\name{get.quantiles}
\alias{get.quantiles}
\alias{get.ind.quantiles}
\title{function to extract quantiles from INLA output}
\usage{
get.quantiles(mylist, quantiles, single)

get.ind.quantiles(outmat, inmat)
}
\arguments{
\item{mylist}{list of matrices of two cols x, y}

\item{quantiles}{vector with the desired quantiles}

\item{single}{NULL or TRUE if only a single node and parameter}

\item{outmat}{matrix where the first col has the desired quantiles. We want to estimate this and out in into the second col}

\item{inmat}{is the actual x,f(x) matrix}
}
\value{
list

matrix
}
\description{
function to get to extract quantiles
}
\section{Functions}{
\itemize{
\item \code{get.ind.quantiles()}: helper function for get.quantiles

}}
\keyword{internal}
