% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abn-internal.R
\name{get.var.types}
\alias{get.var.types}
\title{Create ordered vector with integers denoting the distribution}
\usage{
get.var.types(data.dists = NULL)
}
\arguments{
\item{data.dists}{list specifying each columns distribution type. Names correspond to column names and values must be one of "gaussian", "binomial", "poisson", "multinomial".}
}
\value{
numeric encoding of distribution corresponding to its list element number in \code{data.dists}.
}
\description{
gaussian = 1, binomial = 2, poisson = 3, multinomial = 4
}
\keyword{internal}
