% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{get_link_function}
\alias{get_link_function}
\title{Helper function to determine link function from distribution}
\usage{
get_link_function(distribution)
}
\description{
Helper function to determine link function from distribution
}
\keyword{internal}
