% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitabn_bayes.R
\name{getMargsINLA}
\alias{getMargsINLA}
\title{function to extract marginals from INLA output}
\usage{
getMargsINLA(list.fixed, list.hyper)
}
\arguments{
\item{list.fixed}{list of matrices of two cols x, y}

\item{list.hyper}{list of hyperparameters}
}
\value{
vector
}
\description{
function to extract marginals from INLA output
}
\keyword{internal}
