% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitabn_bayes.R
\name{getModeVector}
\alias{getModeVector}
\title{function to extract the mode from INLA output}
\usage{
getModeVector(list.fixed, list.hyper)
}
\arguments{
\item{list.fixed}{list of matrices of two cols x, y}

\item{list.hyper}{list of hyperparameters}
}
\value{
vector
}
\description{
function to extract the mode from INLA output
}
\keyword{internal}
