% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{irls_binomial_cpp}
\alias{irls_binomial_cpp}
\title{Iterative Reweighed Least Square algorithm for Binomials}
\usage{
irls_binomial_cpp(A, b, maxit, tol)
}
\value{
a list
}
\description{
IRLS to estimate network score of Binomial nodes.
}
\keyword{internal}
