% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{irls_gaussian_cpp}
\alias{irls_gaussian_cpp}
\title{Iterative Reweighed Least Square algorithm for Gaussians}
\usage{
irls_gaussian_cpp(A, b, maxit, tol)
}
\value{
a list
}
\description{
IRLS to estimate network score of Gaussian nodes.
}
\keyword{internal}
