% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{irls_gaussian_cpp_fast}
\alias{irls_gaussian_cpp_fast}
\title{Fast Iterative Reweighed Least Square algorithm for Gaussians}
\usage{
irls_gaussian_cpp_fast(A, b, maxit, tol)
}
\description{
IRLS to estimate network score of Gaussian nodes.
}
\keyword{internal}
