% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{irls_poisson_cpp}
\alias{irls_poisson_cpp}
\title{Iterative Reweighed Least Square algorithm for Poissons}
\usage{
irls_poisson_cpp(A, b, maxit, tol)
}
\value{
a list
}
\description{
IRLS to estimate network score of Poisson nodes.
}
\keyword{internal}
