% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictABN.R
\name{L_binomial}
\alias{L_binomial}
\title{Computing the log-likelihood}
\usage{
L_binomial(y, x, coef, continuous_part)
}
\arguments{
\item{y}{To fill}

\item{x}{To fill}

\item{coef}{To fill}

\item{continuous_part}{To fill}
}
\value{
To fill
}
\description{
Computes the log-likelihood of a binomial variable
}
