% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictABN.R
\name{L_gaussian}
\alias{L_gaussian}
\title{Computing the log-likelihood}
\usage{
L_gaussian(y, x, coef, var, continuous_part)
}
\arguments{
\item{y}{To fill}

\item{x}{To fill}

\item{coef}{To fill}

\item{var}{To fill}

\item{continuous_part}{To fill}
}
\value{
To fill
}
\description{
Computes the log-likelihood of a Gaussian variable
}
