% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abn-toolbox.R
\name{logit}
\alias{logit}
\title{Logit of proportions}
\usage{
logit(x)
}
\arguments{
\item{x}{numeric with values between \code{[0,1]}.}
}
\value{
numeric vector of same length as \code{x}.

numeric vector of same length as \code{x}.
}
\description{
See also the C implementation \code{?abn::logit_cpp()}.
}
\keyword{utilities}
