% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictABN.R
\name{LogL_poisson}
\alias{LogL_poisson}
\title{Compute the log-likelihood of a Poisson variable}
\usage{
LogL_poisson(y, x, coef, continuous_part)
}
\arguments{
\item{y}{The observed count for the Poisson-distributed variable (must be non-negative).}

\item{x}{A value for a parent of the Poisson variable.}

\item{coef}{The coefficient that links \code{y} to \code{x}.}

\item{continuous_part}{The intercept and any additional contributions from other parents of \code{y}.}
}
\value{
The value of the log-likelihood.
}
\description{
This function computes the log-likelihood of a Poisson-distributed variable
}
