% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abn-methods.R
\name{logLik.abnFit}
\alias{logLik.abnFit}
\title{Print logLik of objects of class \code{abnFit}}
\usage{
\method{logLik}{abnFit}(object, digits = 3L, verbose = TRUE, ...)
}
\arguments{
\item{object}{Object of class \code{abnFit}}

\item{digits}{number of digits of the results.}

\item{verbose}{print additional output.}

\item{...}{additional parameters. Not used at the moment.}
}
\value{
prints the logLik of the fitted model.
}
\description{
Print logLik of objects of class \code{abnFit}
}
