% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abn-methods.R
\name{nobs.abnFit}
\alias{nobs.abnFit}
\title{Print number of observations of objects of class \code{abnFit}}
\usage{
\method{nobs}{abnFit}(object, ...)
}
\arguments{
\item{object}{Object of class \code{abnFit}}

\item{...}{additional parameters. Not used at the moment.}
}
\value{
prints the number of observations of the fitted model.
}
\description{
Print number of observations of objects of class \code{abnFit}
}
