% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abn-toolbox.R
\name{odds}
\alias{odds}
\title{Probability to odds}
\usage{
odds(x)
}
\arguments{
\item{x}{numeric vector of probabilities with values between \code{[0,1]}.}
}
\value{
numeric vector of same length as \code{x}.
}
\description{
Probability to odds
}
\keyword{utilities}
