% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pigs.vienna}
\alias{pigs.vienna}
\title{Dataset related to diseases present in `finishing pigs', animals about to enter the human food chain at an abattoir.}
\format{
A data frame with a mixture of 10 discrete variables, each of which is set as a factor, and a grouping variable.
\describe{
\item{PC}{Binary.}
\item{PT}{Binary. }
\item{MS}{Binary.}
\item{HS}{Binary.}
\item{TAIL}{Binary.}
\item{Abscess}{Binary.}
\item{Pyaemia}{Binary.}
\item{EPcat}{Binary.}
\item{PDcat}{Binary.}
\item{plbinary}{Binary.}
\item{batch}{Group variable, corresponding to the 500 different pig producers}
}
}
\usage{
pigs.vienna
}
\description{
The data we consider here comprise of a randomly chosen batch of 50 pigs from
each of 500 randomly chosen pig producers in the UK.
The dataset consists of 25000 observations, 10 binary variables, and a grouping variable.
These are `finishing pigs', animals about to enter the human food chain at an abattoir.
Further description of the data set is present on the vignette.
}
\details{
This dataset was used in an older version of the vignette.
See also the files provided in the package directories
\code{inst/bootstrapping_example} and \code{inst/old_vignette} give a
detailed analysis of the dataset and provide more details for a
bootstrapping example thereof.
}
\references{
Hartnack, S., et al. (2016) "Attitudes of Austrian veterinarians towards euthanasia in small animal practice: impacts of age and gender on views on euthanasia." BMC Veterinary Research 12.1: 26.
}
\keyword{datasets}
\keyword{internal}
