% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abn-methods.R
\name{plot.abnDag}
\alias{plot.abnDag}
\title{Plots DAG from an object of class \code{abnDag}}
\usage{
\method{plot}{abnDag}(x, new = TRUE, ...)
}
\arguments{
\item{x}{Object of class \code{abnDag}}

\item{new}{defaults to \code{TRUE} for opening a new plot.}

\item{...}{additional parameters. Not used at the moment.}
}
\value{
\code{Rgraphviz::plot}
}
\description{
Plots DAG from an object of class \code{abnDag}
}
\examples{
mydag <- createAbnDag(dag = ~a+b|a, data.df = data.frame("a"=1, "b"=1))
plot(mydag)
}
\concept{DAG}
