% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abn-methods.R
\name{plot.abnHeuristic}
\alias{plot.abnHeuristic}
\title{Plot objects of class \code{abnHeuristic}}
\usage{
\method{plot}{abnHeuristic}(x, ...)
}
\arguments{
\item{x}{Object of class \code{abnHeuristic}}

\item{...}{additional parameters. Not used at the moment.}
}
\value{
plot of the scores of the heuristic search.
}
\description{
Plot objects of class \code{abnHeuristic}
}
