% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abn-methods.R
\name{plot.abnHillClimber}
\alias{plot.abnHillClimber}
\title{Plot objects of class \code{abnHillClimber}}
\usage{
\method{plot}{abnHillClimber}(x, ...)
}
\arguments{
\item{x}{Object of class \code{abnHillClimber}}

\item{...}{additional parameters. Not used at the moment.}
}
\value{
plot of the consensus DAG.
}
\description{
Plot objects of class \code{abnHillClimber}
}
