% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abn-methods.R
\name{plot.abnMostprobable}
\alias{plot.abnMostprobable}
\title{Plot objects of class \code{abnMostprobable}}
\usage{
\method{plot}{abnMostprobable}(x, ...)
}
\arguments{
\item{x}{Object of class \code{abnMostprobable}}

\item{...}{additional parameters. Not used at the moment.}
}
\value{
plot of the mostprobable consensus DAG.
}
\description{
Plot objects of class \code{abnMostprobable}
}
