% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictABN.R
\name{plot_Abn}
\alias{plot_Abn}
\title{Plot ABN fitted network}
\usage{
plot_Abn(dag, mydists, node, order)
}
\arguments{
\item{dag}{An adjacency matrix (can be the output of the function mostProbable()).}

\item{mydists}{A list containing the distributions of the nodes of the graph.}

\item{node}{The node of interest.}

\item{order}{up or down indicating whether we predict the node from upstream or downstream.}
}
\value{
A graph object.
}
\description{
This function plots the ABN network with an emphasis on a specific node
}
\examples{
# load a data set
data <- ex1.dag.data 

# define the distributions of the node
mydists <- list(b1="binomial", 
p1="poisson", 
g1="gaussian", 
b2="binomial", 
p2="poisson", 
b3="binomial", 
g2="gaussian", 
b4="binomial", 
b5="binomial", 
g3="gaussian") 

node <- "b4"
order <- "up"
plot_node <- plot_Abn(dag, mydists, node, order)
plot_node
}
