% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictABN.R
\name{plot_workflow}
\alias{plot_workflow}
\title{Plotting the procedure's workflow}
\usage{
plot_workflow(dag, mydists, hypothesis, path = NULL, directory.name = NULL)
}
\arguments{
\item{dag}{An adjacency matrix (can be the output of the function mostProbable()).}

\item{mydists}{A list containing the distributions of the nodes of the graph.}

\item{hypothesis}{Node to predict.}

\item{path}{A valid path to save the plots}

\item{directory.name}{The name of the directory that will be created to save the plots.}
}
\value{
An animated gif that represents the procedure's workflow.
}
\description{
Plots and saves an animated gif that represents the procedure's workflow.
}
\examples{
# load a data set
data <- ex1.dag.data 

# define the distributions of the node
mydists <- list(b1="binomial", 
p1="poisson", 
g1="gaussian", 
b2="binomial", 
p2="poisson", 
b3="binomial", 
g2="gaussian", 
b4="binomial", 
b5="binomial", 
g3="gaussian") 

# infer the graph using ABN
max.par <- 4 # set the same max parents for all nodes
mycache <- buildScoreCache(data.df = data, 
                          data.dists = mydists,
                          method = "bayes",max.parents = max.par) 
mp.dag <- mostProbable(score.cache = mycache)
dag <- mp.dag$dag

hypothesis <- "g2"
plot_workflow(dag, mydists, hypothesis)
}
