% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictABN.R
\name{plotPosteriorDistrib}
\alias{plotPosteriorDistrib}
\title{Plot the posterior distribution}
\usage{
plotPosteriorDistrib(predictions, hypothesis, mydists)
}
\arguments{
\item{predictions}{To fill.}

\item{hypothesis}{To fill.}

\item{mydists}{To fill.}
}
\value{
None
}
\description{
This functions plots the posterior distribution
}
