% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictABN.R
\name{predict_node_from_children_binomial}
\alias{predict_node_from_children_binomial}
\title{Performing inference with ABN (downstream inference)}
\usage{
predict_node_from_children_binomial(
  data,
  mydists,
  myfit,
  node,
  evidence,
  children,
  parents,
  predictions
)
}
\arguments{
\item{data}{A data frame containing the data (samples in rows, variables in columns).}

\item{mydists}{A list containing the distributions of the nodes of the graph.}

\item{myfit}{Parameters of the network (can be the output of the function fitAbn()).}

\item{node}{Temporary node to predict.}

\item{evidence}{Known nodes that are used to predict the hypothesis.}

\item{children}{The children of the node to predict.}

\item{parents}{The children's parents of the node to predict.}

\item{predictions}{The estimated predictions of the upstream nodes.}
}
\value{
The predicted distribution of the node of interest.
}
\description{
Main function to predict the distribution of a binomial node given its children and its children's parents
}
\examples{
# load a data set
data <- ex1.dag.data 

# define the distributions of the node
mydists <- list(b1="binomial", 
p1="poisson", 
g1="gaussian", 
b2="binomial", 
p2="poisson", 
b3="binomial", 
g2="gaussian", 
b4="binomial", 
b5="binomial", 
g3="gaussian") 

# infer the graph using ABN
max.par <- 4 # set the same max parents for all nodes
mycache <- buildScoreCache(data.df = data, 
                          data.dists = mydists,
                          method = "bayes",max.parents = max.par) 
mp.dag <- mostProbable(score.cache = mycache)
dag <- mp.dag$dag
graph <- graph_from_adjacency_matrix(t(dag))

# infer the parameters of the network
myfit <- fitAbn(object = mp.dag)
myfit <- myfit$modes

node <- "g2"
children <- find_children(graph, node)[1]
parents <- find_parents(graph, children)
evidence <- list("p1" = 3, "b5" = "y")

predictions <- predict_node_from_children_gaussian(data, mydists, myfit, node, evidence, children, parents)
predictions  # the predicted distribution of g2
}
