% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictABN.R
\name{predict_node_from_children_poisson}
\alias{predict_node_from_children_poisson}
\title{Perform downstream inference with ABN}
\usage{
predict_node_from_children_poisson(
  data,
  mydists,
  myfit,
  node,
  evidence,
  child,
  parents,
  predictions
)
}
\arguments{
\item{data}{A data frame containing the data (samples in rows, variables in columns).}

\item{mydists}{A list containing the distributions of the nodes of the graph.}

\item{myfit}{Parameters of the network (can be the output of the function fitAbn()).}

\item{node}{Temporary node to predict.}

\item{evidence}{Known nodes that are used to predict the hypothesis.}

\item{parents}{The children's parents of the node to predict.}

\item{predictions}{The estimated predictions of the downstream nodes (must contain at least a first prediction of the node to predict if the child and its parents are evidence).}

\item{children}{The children of the node to predict.}
}
\value{
The predicted distribution of the node of interest.
}
\description{
Main function to predict the distribution of a node given one of its Poisson child and its parents
}
\examples{
# load a data set
data <- ex1.dag.data 

# define the distributions of the node
mydists <- list(b1="binomial", 
p1="poisson", 
g1="gaussian", 
b2="binomial", 
p2="poisson", 
b3="binomial", 
g2="gaussian", 
b4="binomial", 
b5="binomial", 
g3="gaussian") 

# infer the graph using ABN
max.par <- 4 # set the same max parents for all nodes
mycache <- buildScoreCache(data.df = data, 
                          data.dists = mydists,
                          method = "bayes",max.parents = max.par) 
mp.dag <- mostProbable(score.cache = mycache)
dag <- mp.dag$dag
graph <- graph_from_adjacency_matrix(t(dag))

# infer the parameters of the network
myfit <- fitAbn(object = mp.dag)
myfit <- myfit$modes

node <- "b1"
child <- find_children(graph, node)[1]
parents <- find_parents(graph, child)
evidence <- list("p1" = 3, "p2" = 4)
evidence <- check_evidence(data, mydists, hypothesis = node,evidence) # check the format of the evidence

predictions <- list("b1" = c(0.5,0.5)) # a first estimate of the node b1

predictions <- predict_node_from_children_poisson(data, mydists, myfit, node, evidence, child, parents, predictions)
predictions  # the predicted distribution of g2
}
