% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictABN.R
\name{predict_node_from_parent_poisson}
\alias{predict_node_from_parent_poisson}
\title{Perform upstream inference with ABN}
\usage{
predict_node_from_parent_poisson(
  data,
  mydists,
  myfit,
  node,
  evidence,
  parents,
  predictions = NULL
)
}
\arguments{
\item{data}{A data frame containing the data (samples in rows, variables in columns).}

\item{mydists}{A list containing the distributions of the nodes of the graph.}

\item{myfit}{Parameters of the network (can be the output of the function fitAbn()).}

\item{node}{Temporary node to predict.}

\item{evidence}{Known nodes that are used to predict the hypothesis.}

\item{parents}{The parents of the node to predict.}

\item{predictions}{The estimated predictions of the upstream nodes  (can be empty if the parents of the node to predict are evidence).}
}
\value{
The predicted distribution of the node of interest.
}
\description{
Main function to predict the distribution of a Poisson node given its parents only
}
\examples{
# load a data set
data <- ex1.dag.data 

# define the distributions of the node
mydists <- list(b1="binomial", 
p1="poisson", 
g1="gaussian", 
b2="binomial", 
p2="poisson", 
b3="binomial", 
g2="gaussian", 
b4="binomial", 
b5="binomial", 
g3="gaussian") 

# infer the graph using ABN
max.par <- 4 # set the same max parents for all nodes
mycache <- buildScoreCache(data.df = data, 
                          data.dists = mydists,
                          method = "bayes",max.parents = max.par) 
mp.dag <- mostProbable(score.cache = mycache)
dag <- mp.dag$dag
graph <- graph_from_adjacency_matrix(t(dag))

# infer the parameters of the network
myfit <- fitAbn(object = mp.dag)
myfit <- myfit$modes

node <- "p2"
parents <- find_parents(graph,node)

evidence <- list("b1" = "y", "p1" = 3)

predictions <- predict_node_from_parent_poisson(data, mydists, myfit, node, evidence, parents)
predictions  # the predicted distribution of p1
}
