% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictABN.R
\name{predict_root}
\alias{predict_root}
\title{Predict a root node}
\usage{
predict_root(data, mydists, node)
}
\arguments{
\item{data}{A data frame containing the data (samples in rows, variables in columns).}

\item{mydists}{A list containing the distributions of the nodes of the graph.}

\item{node}{Root node to predict.}
}
\value{
The predicted distribution of the root node.
}
\description{
Predict a root node in the graph
}
\examples{
# load a data set
data <- ex1.dag.data 

# define the distributions of the node
mydists <- list(b1="binomial", 
p1="poisson", 
g1="gaussian", 
b2="binomial", 
p2="poisson", 
b3="binomial", 
g2="gaussian", 
b4="binomial", 
b5="binomial", 
g3="gaussian") 

prediction <- predict_root(data, mydists, node = "g3")
}
