% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictABN.R
\name{predictABN}
\alias{predictABN}
\title{Perform inference with ABN}
\usage{
predictABN(data, mydists, dag, myfit, hypothesis, evidence, plot = FALSE)
}
\arguments{
\item{data}{A data frame containing the data (samples in rows, variables in columns).}

\item{mydists}{A list containing the distributions of the nodes of the graph.}

\item{dag}{An adjacency matrix (can be the output of the function mostProbable()).}

\item{myfit}{Parameters of the network (can be the output of the function fitAbn()).}

\item{hypothesis}{Node to predict.}

\item{evidence}{Known nodes that are used to predict the hypothesis.}

\item{plot}{TRUE/FALSE to indicate if the predicted distribution has to be plotted.}
}
\value{
A list containing the predicted distribution of the hypothesis and predicted distributions of the upstream nodes.
}
\description{
Main function to predict the distribution of a node in a fitted ABN graph
}
\examples{
# load a data set
data <- ex1.dag.data 

# define the distributions of the node
mydists <- list(b1="binomial", 
p1="poisson", 
g1="gaussian", 
b2="binomial", 
p2="poisson", 
b3="binomial", 
g2="gaussian", 
b4="binomial", 
b5="binomial", 
g3="gaussian") 

# infer the graph using ABN
max.par <- 4 # set the same max parents for all nodes
mycache <- buildScoreCache(data.df = data, 
                          data.dists = mydists,
                          method = "bayes",max.parents = max.par) 
mp.dag <- mostProbable(score.cache = mycache)
dag <- mp.dag$dag

# infer the parameters of the network
myfit <- fitAbn(object = mp.dag)
myfit <- myfit$modes

hypothesis <- "g2"

evidence <- list("p1" = 3)

predictions <- predictABN(data, mydists, dag, myfit, hypothesis, evidence)
str(predictions)
predictions$prediction_hypothesis # the predicted distribution of g2
}
