% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abn-methods.R
\name{print.abnDag}
\alias{print.abnDag}
\title{Print objects of class \code{abnDag}}
\usage{
\method{print}{abnDag}(x, digits = 3L, ...)
}
\arguments{
\item{x}{Object of class \code{abnDag}}

\item{digits}{number of digits of the adjacency matrix.}

\item{...}{additional parameters. Not used at the moment.}
}
\value{
outputs adjacency matrix and statement of the class of \code{x}.
}
\description{
Print objects of class \code{abnDag}
}
\examples{
mydag <- createAbnDag(dag = ~a+b|a, data.df = data.frame("a"=1, "b"=1))
print(mydag)
}
\concept{DAG}
