% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abn-methods.R
\name{print.abnFit}
\alias{print.abnFit}
\title{Print objects of class \code{abnFit}}
\usage{
\method{print}{abnFit}(x, digits = 3L, ...)
}
\arguments{
\item{x}{Object of class \code{abnFit}}

\item{digits}{number of digits of the results.}

\item{...}{additional parameters. Not used at the moment.}
}
\value{
prints the parameters of the fitted model.
}
\description{
Print objects of class \code{abnFit}
}
