% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abn-methods.R
\name{print.abnHeuristic}
\alias{print.abnHeuristic}
\title{Print objects of class \code{abnHeuristic}}
\usage{
\method{print}{abnHeuristic}(x, digits = 2L, ...)
}
\arguments{
\item{x}{Object of class \code{abnHeuristic}}

\item{digits}{number of digits of the results.}

\item{...}{additional parameters. Not used at the moment.}
}
\value{
prints the best score found and the distribution of the scores.
}
\description{
Print objects of class \code{abnHeuristic}
}
