% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abn-methods.R
\name{print.abnHillClimber}
\alias{print.abnHillClimber}
\title{Print objects of class \code{abnHillClimber}}
\usage{
\method{print}{abnHillClimber}(x, digits = 3L, ...)
}
\arguments{
\item{x}{Object of class \code{abnHillClimber}}

\item{digits}{number of digits of the results.}

\item{...}{additional parameters. Not used at the moment.}
}
\value{
prints the consensus DAG and the class of the object.
}
\description{
Print objects of class \code{abnHillClimber}
}
