% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abn-methods.R
\name{print.abnMostprobable}
\alias{print.abnMostprobable}
\title{Print objects of class \code{abnMostprobable}}
\usage{
\method{print}{abnMostprobable}(x, digits = 3L, ...)
}
\arguments{
\item{x}{Object of class \code{abnMostprobable}}

\item{digits}{number of digits of the results.}

\item{...}{additional parameters. Not used at the moment.}
}
\value{
prints the mostprobable consensus DAG.
}
\description{
Print objects of class \code{abnMostprobable}
}
