% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictABN.R
\name{prior_binomial}
\alias{prior_binomial}
\title{Compute prior probability for a binomial prior distribution}
\usage{
prior_binomial(x, p)
}
\arguments{
\item{x}{The observed outcome (0 or 1).}

\item{p}{The probability of sucess (between 0 and 1).}
}
\value{
The probability of observing x given sucess probability p.
}
\description{
This function calculates the probability mass function of a binomial distribution
}
