% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictABN.R
\name{prior_gaussian}
\alias{prior_gaussian}
\title{Computing the prior}
\usage{
prior_gaussian(x, mu, sigma2)
}
\arguments{
\item{x}{To fill}

\item{mu}{To fill}

\item{sigma2}{To fill}
}
\value{
To fill
}
\description{
Computes the prior of a Gaussian variable
}
