% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictABN.R
\name{prior_poisson}
\alias{prior_poisson}
\title{Compute Prior Probability for a Poisson Distribution}
\usage{
prior_poisson(x, lambda)
}
\arguments{
\item{x}{The observed count (number of events, must be non-negative).}

\item{lambda}{The expected number of events (rate parameter, must be positive).}
}
\value{
The probability of observing \code{x} events given rate \code{lambda}.
}
\description{
This function calculates the probability mass function of a Poisson distribution
}
