% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{reconstruct_abnfit_bayes}
\alias{reconstruct_abnfit_bayes}
\title{Reconstruct abnFit object for Bayesian method from JSON}
\usage{
reconstruct_abnfit_bayes(json_list)
}
\description{
Reconstruct abnFit object for Bayesian method from JSON
}
\keyword{internal}
