% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{reconstruct_abnfit_mle}
\alias{reconstruct_abnfit_mle}
\title{Reconstruct abnFit object for MLE method from JSON}
\usage{
reconstruct_abnfit_mle(json_list)
}
\description{
Reconstruct abnFit object for MLE method from JSON
}
\keyword{internal}
