

\name{sim10var}
\alias{sim10var}
\non_function{}
\title{simulated dataset from a 10 variable discrete DAG}
\description{
 1000 observations simulated from a DAG. 
}
\format{A data frame with discrete variables.
  \describe{
    \item{var1}{Binary. }
    \item{var2}{Binary. }
    \item{var3}{Binary. }
    \item{var4}{Binary. }
    \item{var5}{Binary. }
    \item{var6}{Binary. }
    \item{var7}{Binary. }
    \item{var8}{Binary. }
    \item{var9}{Binary. }
    \item{var10}{Binary. }
    }
 }

\keyword{datasets}

