

\name{sim10varadd}
\alias{sim10varadd}
\non_function{}
\title{simulated dataset from a 10 variable additive discrete DAG}
\description{
 1000 observations simulated from a DAG. 
}
\format{A data frame with discrete variables.
  \describe{
    \item{X1}{Binary. }
    \item{X2}{Binary. }
    \item{X3}{Binary. }
    \item{X4}{Binary. }
    \item{X5}{Binary. }
    \item{X6}{Binary. }
    \item{X7}{Binary. }
    \item{X8}{Binary. }
    \item{X9}{Binary. }
    \item{X10}{Binary. }
    }
 }

\keyword{datasets}

