% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abn-toolbox.R
\name{skewness}
\alias{skewness}
\title{Computes skewness of a distribution}
\usage{
skewness(x)
}
\arguments{
\item{x}{a numeric vector}
}
\value{
integer
}
\description{
Computes skewness of a distribution
}
\keyword{utilities}
