% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitabn_bayes.R
\name{std.area.under.grid}
\alias{std.area.under.grid}
\title{Standard Area Under the Marginal}
\usage{
std.area.under.grid(mylist, single)
}
\arguments{
\item{mylist}{list of matrices of two cols x, y}

\item{single}{NULL or TRUE if only a single node and parameter}
}
\value{
list
}
\description{
function to get std. are under marginal to exactly unity.
It should be very close to unity but in some cases due to numerical accuracy
differences (since each point is a separate estimate) this might be a little adrift
turn this option off to see how reliable the original estimation is
}
\keyword{internal}
