% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abn-methods.R
\name{summary.abnDag}
\alias{summary.abnDag}
\title{Prints summary statistics from an object of class \code{abnDag}}
\usage{
\method{summary}{abnDag}(object, ...)
}
\arguments{
\item{object}{an object of class \code{abnLearned}, \code{abnFit}.
Alternatively, a matrix or a formula statement defining the network structure,
a directed acyclic graph (DAG).
Note that row names must be set up or given in \code{node.names}.}

\item{...}{additional parameters. Not used at the moment.}
}
\value{
List with summary statistics of the DAG.
}
\description{
Prints summary statistics from an object of class \code{abnDag}
}
\examples{
mydag <- createAbnDag(dag = ~a+b|a, data.df = data.frame("a"=1, "b"=1))
summary(mydag)
}
\concept{DAG}
