% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abn-methods.R
\name{summary.abnFit}
\alias{summary.abnFit}
\title{Print summary of objects of class \code{abnFit}}
\usage{
\method{summary}{abnFit}(object, digits = 3L, ...)
}
\arguments{
\item{object}{Object of class \code{abnFit}}

\item{digits}{number of digits of the results.}

\item{...}{additional parameters. Not used at the moment.}
}
\value{
prints summary statistics of the fitted model.
}
\description{
Print summary of objects of class \code{abnFit}
}
