% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{validate_abnfit_object}
\alias{validate_abnfit_object}
\title{Validate abnFit object meets class requirements}
\usage{
validate_abnfit_object(object)
}
\description{
Validate abnFit object meets class requirements
}
\keyword{internal}
