% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import.R
\name{validate_json_structure}
\alias{validate_json_structure}
\title{Validate JSON structure for abnFit import}
\usage{
validate_json_structure(json_list)
}
\description{
Validate JSON structure for abnFit import
}
\keyword{internal}
