% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{var33}
\alias{var33}
\title{simulated dataset from a DAG comprising of 33 variables}
\format{
A data frame with a mixture of discrete variables each of which is
set as a factor and continuous variables.
Joint distribution structure used to generate the data.
\describe{
\item{v1}{Binary, independent. }
\item{v2}{Gaussian, conditionally dependent upon v1. }
\item{v3}{Binary, independent. }
\item{v4}{Binary, conditionally dependent upon v3. }
\item{v5}{Gaussian, conditionally dependent upon v6. }
\item{v6}{Binary, conditionally dependent upon v4 and v7. }
\item{v7}{Gaussian,  conditionally dependent upon v8. }
\item{v8}{Gaussian,  conditionally dependent upon v9. }
\item{v9}{Binary, conditionally dependent upon v10. }
\item{v10}{Binary, independent. }
\item{v11}{Binary,  conditionally dependent upon v10, v12 and v19. }
\item{v12}{Binary, independent.}
\item{v13}{Gaussian, independent.}
\item{v14}{Gaussian,  conditionally dependent upon v13. }
\item{v15}{Binary, conditionally dependent upon v14 and v21. }
\item{v16}{Gaussian, independent. }
\item{v17}{Gaussian, conditionally dependent upon v16 and v20. }
\item{v18}{Binary,  conditionally dependent upon v20. }
\item{v19}{Binary,  conditionally dependent upon v20. }
\item{v20}{Binary, independent. }
\item{v21}{Binary, conditionally dependent upon v20. }
\item{v22}{Gaussian, conditionally dependent upon v21. }
\item{v23}{Gaussian, conditionally dependent upon v21. }
\item{v24}{Gaussian, conditionally dependent upon v23. }
\item{v25}{Gaussian, conditionally dependent upon v23 and v26. }
\item{v26}{Binary, conditionally dependent upon v20. }
\item{v27}{Binary, independent. }
\item{v28}{Binary, conditionally dependent upon v27, v29 and v31. }
\item{v29}{Gaussian, independent. }
\item{v30}{Gaussian,  conditionally dependent upon v29. }
\item{v31}{Gaussian, independent. }
\item{v32}{Binary,  conditionally dependent upon v21, v29 and v31. }
\item{v33}{Gaussian,  conditionally dependent upon v31. }
}
}
\usage{
var33
}
\description{
250 observations simulated from a DAG with 17 binary variables and 16 continuous.
A DAG of this data features in the vignette.
Note that the conditional dependence relations given are those in the population
and may differ in the realization of 250 observations.
}
\examples{
## Constructing the DAG of the dataset:
dag33 <- matrix(0, 33, 33)
dag33[2,1] <- 1
dag33[4,3] <- 1
dag33[6,4] <- 1; dag33[6,7] <- 1
dag33[5,6] <- 1
dag33[7,8] <- 1
dag33[8,9] <- 1
dag33[9,10] <- 1
dag33[11,10] <- 1; dag33[11,12] <- 1; dag33[11,19] <- 1;
dag33[14,13] <- 1
dag33[17,16] <- 1; dag33[17,20] <- 1
dag33[15,14] <- 1; dag33[15,21] <- 1
dag33[18,20] <- 1
dag33[19,20] <- 1
dag33[21,20] <- 1
dag33[22,21] <- 1
dag33[23,21] <- 1
dag33[24,23] <- 1
dag33[25,23] <- 1; dag33[25,26] <- 1
dag33[26,20] <- 1
dag33[33,31] <- 1
dag33[33,31] <- 1
dag33[32,21] <- 1; dag33[32,31] <- 1; dag33[32,29] <- 1
dag33[30,29] <- 1
dag33[28,27] <- 1; dag33[28,29] <- 1; dag33[28,31] <- 1

}
\keyword{datasets}
\keyword{internal}
