% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictABN.R
\name{createAnimation}
\alias{createAnimation}
\title{Create an animated gif}
\usage{
createAnimation(path = NULL)
}
\arguments{
\item{path}{A valid path to a repository that contains the png to combine in a gif.}
}
\value{
None
}
\description{
This function creates an animated gif file based on png files
}
