% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictABN.R
\name{createDirectory}
\alias{createDirectory}
\title{Creating a directory}
\usage{
createDirectory(path = NULL, directory.name = NULL)
}
\arguments{
\item{path}{A valid path.}

\item{directory.name}{The name of the directory that will be created in this path.}
}
\value{
A path corresponding to this created repository
}
\description{
Creates a directory situated in a particular path
}
\examples{
path <- createDirectory(path = NULL, directory.name = "test")
}
