\name{ex3.dag.data}
     \docType{data}
     \alias{ex3.dag.data}
     
     \title{Data set for use with abn library examples}
     \description{1000 observations across with 13 binary variables and one grouping variable. Real (anonymised) data of unknown structure. 
     }
     \usage{ex3.dag.data} 
\format{A data frame, binary variables are factors
     \describe{
     \item{b1}{binary}
     \item{b2}{binary}
\item{b3}{binary}
\item{b4}{binary}
\item{b5}{binary}
\item{b6}{binary}
\item{b7}{binary}
\item{b8}{binary}
\item{b9}{binary}
\item{b10}{binary}
\item{b11}{binary}
\item{b12}{binary}
\item{b13}{binary}
\item{group}{binary, defines sampling groups}
     }}


\keyword{datasets}
