% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictABN.R
\name{EvaluatePerf}
\alias{EvaluatePerf}
\title{Evaluating the performances}
\usage{
EvaluatePerf(observations, predictions, distribution, compare.distrib = FALSE)
}
\arguments{
\item{observations}{To fill}

\item{predictions}{To fill}

\item{distribution}{To fill}

\item{compare.distrib}{To fill}
}
\value{
To fill
}
\description{
Evaluates the performances of the predictions
}
