% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_abnFit_bayes}
\alias{export_abnFit_bayes}
\title{Export abnFit object fitted with Bayesian methods}
\usage{
export_abnFit_bayes(object, format, include_network, ...)
}
\arguments{
\item{object}{An object of class abnFit}

\item{format}{The export format, currently only "json" is supported.}

\item{include_network}{Whether to include network structure}

\item{...}{Additional export options}
}
\value{
This function does not return a value. It raises an error.
}
\description{
Export abnFit object fitted with Bayesian methods
}
\details{
This function is a placeholder for exporting abnFit objects fitted using Bayesian methods.
Currently, it raises an error indicating that export for Bayesian models is not implemented.
}
\keyword{internal}
