% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{export_abnFit_mle_arcs}
\alias{export_abnFit_mle_arcs}
\title{Export arc information from abnFit objects fitted with MLE}
\usage{
export_abnFit_mle_arcs(object, ...)
}
\arguments{
\item{object}{An object of class abnFit}

\item{...}{Additional export options}
}
\value{
An array containing arc details: source_variable_id and target_variable_id for each arc.
}
\description{
Export arc information from abnFit objects fitted with MLE
}
\details{
This function extracts arc information from abnFit objects fitted using MLE.
It retrieves the source and target nodes for each arc in the directed acyclic graph (DAG).
Currently, frequency, significance, and constraint information are not included in the export.
}
\keyword{internal}
