% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\name{extract_states_from_data}
\alias{extract_states_from_data}
\title{Extract states for categorical variables from data}
\usage{
extract_states_from_data(object, node_id)
}
\description{
Extract states for categorical variables from data
}
\keyword{internal}
