% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitabn_bayes.R
\name{getMSEfromModes}
\alias{getMSEfromModes}
\title{Extract Standard Deviations from all Gaussian Nodes}
\usage{
getMSEfromModes(modes, dists)
}
\arguments{
\item{modes}{list of modes.}

\item{dists}{list of distributions.}
}
\value{
named numeric vector. Names correspond to node name. Value to standard deviations.
}
\description{
Extract Standard Deviations from all Gaussian Nodes
}
