% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictABN.R
\name{L_binomial}
\alias{L_binomial}
\title{Compute the likelihood of a binomial variable}
\usage{
L_binomial(y, x, coef, continuous_part)
}
\arguments{
\item{y}{The observed binary outcome (numeric 0 or 1)}

\item{x}{A value for a parent of the binomial variable.}

\item{coef}{The coefficient that links \code{y} to \code{x}.}

\item{continuous_part}{The rest of the equation (intercept and links between the parents of the child and the child).}
}
\value{
The value of the likelihood.
}
\description{
This function computes the likelihood of a binomial-distributed variable
}
