% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictABN.R
\name{L_gaussian}
\alias{L_gaussian}
\title{Compute the likelihood of a Gaussian variable}
\usage{
L_gaussian(y, x, coef, var, continuous_part)
}
\arguments{
\item{y}{The observed value for the Gaussian-distributed variable.}

\item{x}{A value for a parent of the Gaussian variable.}

\item{coef}{The coefficient that links \code{y} to \code{x}.}

\item{var}{The variance of the Gaussian variable.}

\item{continuous_part}{The rest of the equation (intercept and links between the parents of the child and the child).}
}
\value{
The value of the likelihood.
}
\description{
This function computes the likelihood of a Gaussian-distributed variable
}
