% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{logit_cpp}
\alias{logit_cpp}
\alias{expit_cpp}
\title{logit and logit functions}
\usage{
logit_cpp(x)

expit_cpp(x)
}
\arguments{
\item{x}{a numeric vector}
}
\description{
transform \code{x} either via the logit, or expit.
}
