% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitabn_bayes.R
\name{modes2coefs}
\alias{modes2coefs}
\title{Convert modes to fitAbn.mle$coefs structure}
\usage{
modes2coefs(modes)
}
\arguments{
\item{modes}{list of modes.}
}
\value{
list of matrix arrays.
}
\description{
Convert modes to fitAbn.mle$coefs structure
}
