%                               -*- Mode: Rd -*- 
% or.Rd --- 
% Author          : Gilles Kratzer
% Created On      : 18.07.2018
% Last Modified By: 
% Last Modified On: 
% Update Count    : 
% Status          : Unknown, Use with caution!
%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%

\name{or}
\alias{or}

\encoding{latin1}
%- Also NEED an `\alias' for EACH other topic documented here.

\title{Odds Ratio from a Table}

\description{Compute the odds ratio from a table or a matrix.}

\usage{
or(x)
}

\arguments{
  \item{x}{a 2x2 table or matrix.}
}

\details{
Compute the odds ratio from a table or a matrix.
}
\value{
A real value.
}

\author{Gilles Kratzer}

\keyword{utilities}
