% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abn-toolbox.R
\name{or}
\alias{or}
\title{Odds Ratio from a matrix}
\usage{
or(x)
}
\arguments{
\item{x}{a 2x2 table or matrix.}
}
\value{
A real value.
}
\description{
Compute the odds ratio from a contingency table or a matrix.
}
\keyword{utilities}
