% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abn-methods.R
\name{plot.abnDag}
\alias{plot.abnDag}
\title{Plots DAG from an object of class \code{abnDag}}
\usage{
\method{plot}{abnDag}(x, ...)
}
\arguments{
\item{x}{Object of class \code{abnDag}}

\item{...}{additional parameters. Not used at the moment.}
}
\value{
\code{Rgraphviz::plot}
}
\description{
Plots DAG from an object of class \code{abnDag}
}
\examples{
mydag <- createAbnDag(dag = ~a+b|a,
                      data.df = data.frame("a"=1, "b"=1),
                      data.dists = list(a="binomial", b="gaussian"))
plot(mydag)
}
\concept{DAG}
