% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abn-methods.R
\name{plot.abnFit}
\alias{plot.abnFit}
\title{Plot objects of class \code{abnFit}}
\usage{
\method{plot}{abnFit}(x, which = "abnFit", ...)
}
\arguments{
\item{x}{Object of class \code{abnFit}}

\item{which}{defaults to "abnFit".}

\item{...}{additional parameters. Not used at the moment.}
}
\description{
Plot objects of class \code{abnFit}
}
