% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abn-methods.R
\name{plot.abnHillClimber}
\alias{plot.abnHillClimber}
\title{Plot objects of class \code{abnHillClimber}}
\usage{
\method{plot}{abnHillClimber}(x, new = TRUE, ...)
}
\arguments{
\item{x}{Object of class \code{abnHillClimber}}

\item{new}{defaults to \code{TRUE} for opening a new plot.}

\item{...}{additional parameters. Not used at the moment.}
}
\description{
Plot objects of class \code{abnHillClimber}
}
