% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abn-methods.R
\name{plot.abnMostprobable}
\alias{plot.abnMostprobable}
\title{Plot objects of class \code{abnMostprobable}}
\usage{
\method{plot}{abnMostprobable}(x, new = TRUE, ...)
}
\arguments{
\item{x}{Object of class \code{abnMostprobable}}

\item{new}{defaults to \code{TRUE} for opening a new plot.}

\item{...}{additional parameters. Not used at the moment.}
}
\description{
Plot objects of class \code{abnMostprobable}
}
