% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictABN.R
\name{plotPosteriorDistrib}
\alias{plotPosteriorDistrib}
\title{Plotting the posterior distribution}
\usage{
plotPosteriorDistrib(predictions, hypothesis, mydists, path)
}
\arguments{
\item{predictions}{To fill.}

\item{hypothesis}{To fill.}

\item{mydists}{To fill.}

\item{path}{To fill.}
}
\value{
None
}
\description{
Plots the posterior distribution
}
