% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictABN.R
\name{predict_node_from_parent}
\alias{predict_node_from_parent}
\title{Performing inference with ABN (upstream inference)}
\usage{
predict_node_from_parent(
  data,
  mydists,
  graph,
  myfit,
  node,
  evidence,
  predictions = evidence
)
}
\arguments{
\item{data}{A data frame containing the data (samples in rows, variables in columns).}

\item{mydists}{A list containing the distributions of the nodes of the graph.}

\item{graph}{A dag (igraph object).}

\item{myfit}{Parameters of the network (can be the output of the function fitAbn()).}

\item{node}{Temporary node to predict.}

\item{evidence}{Known nodes that are used to predict the hypothesis.}

\item{predictions}{The estimated predictions of the upstream nodes.}
}
\value{
The predicted distribution of the node of interest.
}
\description{
Main function to predict the distribution of one node given its parents only
}
\examples{
# load a data set
data <- ex1.dag.data 

# define the distributions of the node
mydists <- list(b1="binomial", 
p1="poisson", 
g1="gaussian", 
b2="binomial", 
p2="poisson", 
b3="binomial", 
g2="gaussian", 
b4="binomial", 
b5="binomial", 
g3="gaussian") 

# infer the graph using ABN
max.par <- 4 # set the same max parents for all nodes
mycache <- buildScoreCache(data.df = data, 
                          data.dists = mydists,
                          method = "bayes",max.parents = max.par) 
mp.dag <- mostProbable(score.cache = mycache)
dag <- mp.dag$dag
graph <- graph_from_adjacency_matrix(t(dag))

# infer the parameters of the network
myfit <- fitAbn(object = mp.dag)
myfit <- myfit$modes

node <- "g2"

evidence <- list("p1" = 3, "g3" = 5)

predictions <- predict_node_from_parent(data, mydists, graph, myfit, node, evidence)
predictions  # the predicted distribution of g2
}
