% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictABN.R
\name{prior_binomial}
\alias{prior_binomial}
\title{Computing the prior}
\usage{
prior_binomial(x, p)
}
\arguments{
\item{x}{To fill}

\item{p}{To fill}
}
\value{
To fill
}
\description{
Computes the prior of a binomial variable
}
