% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictABN.R
\name{prior_gaussian}
\alias{prior_gaussian}
\title{Compute Gaussian prior density}
\usage{
prior_gaussian(x, mu, sigma2)
}
\arguments{
\item{x}{The value at which the density is evaluated.}

\item{mu}{The expected value of the Gaussian distribution.}

\item{sigma2}{The variance of the Gaussian distribution.}
}
\value{
The density of the normal distribution at x.
}
\description{
This function calculates the density of a Gaussian distribution
}
