% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PredictABN.R
\name{prior_poisson}
\alias{prior_poisson}
\title{Computing the prior}
\usage{
prior_poisson(x, lambda)
}
\arguments{
\item{x}{To fill}

\item{lambda}{To fill}
}
\value{
To fill
}
\description{
Computes the prior of a Poisson variable
}
