% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abn-internal.R
\name{strsplits}
\alias{strsplits}
\title{Recursive string splitting}
\usage{
strsplits(x, splits, ...)
}
\description{
Internal function that call multiple times strsplit() and remove space
}
\keyword{internal}
