% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/abn-methods.R
\name{summary.abnMostprobable}
\alias{summary.abnMostprobable}
\title{Print summary of objects of class \code{abnMostprobable}}
\usage{
\method{summary}{abnMostprobable}(object, ...)
}
\arguments{
\item{object}{Object of class \code{abnMostprobable}}

\item{...}{additional parameters. Not used at the moment.}
}
\value{
prints the mostprobable consensus DAG and the number of observations used to calculate it.
}
\description{
Print summary of objects of class \code{abnMostprobable}
}
